pyproject.toml
[tool.poetry]
name = "api-comunidade-iv"
version = "0.0.1"
description = ""
authors = ["Hylan Silva <hylansilva@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
